+++ Sources/StreamChatUI/Appearance+Images.swift
@@ -205,6 +205,12 @@ public extension Appearance {
             }
             set { _fileAttachmentActionIcons = newValue }
         }
+        
+        public var camera: UIImage = UIImage(named: "camera", in: .streamChatUI)!
+        public var bigPlay: UIImage = UIImage(named: "play_big", in: .streamChatUI)!
+        
+        public var play: UIImage = UIImage(named: "play", in: .streamChatUI)!
+        public var pause: UIImage = UIImage(named: "pause", in: .streamChatUI)!
 
         // MARK: - CommandIcons