+++ CHANGELOG.md
@@ -19,15 +19,31 @@ _June 23, 2021_
 - `LinkPreviewViewDelegate` methods updated to have optional index path
 - `scrollToLatestMessageButton` type changed from `UIButton` to `_ScrollToLatestMessageButton<ExtraData>`
 - `UITableView` is now used instead of `UICollectionView` to display the message list [#1219](https://github.com/GetStream/stream-chat-swift/pull/1219) 
+- `ChatMessageImageGallery` renamed to `ChatMessageGalleryView`, updated to show any content
+- `ImageGalleryVC` renamed to `GalleryVC`
+- `ImagePreviewable` renamed to `GalleryItemPreview`, updated to expose `AttachmentId` only
+- `GalleryContentViewDelegate` methods are renamed to work not only for image attachment but for any
+- `selectedAttachmentType` removed from `ComposerVC`
+- `imagePickerVC` renamed to `mediaPickerVC` in `ComposerVC`
 
 ### ✅ Added
+- Video attachments support:
+ - `VideoAttachmentPayload` type is introduced, video attachments are exposed on `ChatMessage`
+ - `VideoAttachmentComposerView` component is added to displaying video thumbnails in `ComposerVC`
+ - `VideoAttachmentCellView` displaying video previews in `ChatMessageImageGallery`
+ - `VideoCollectionViewCell` displaying videos in `GalleryVC`
+ - `VideoPlaybackControlView` used to take actions on the playing video in `GalleryVC`
+ - `VideoPreviewLoader` loading video thumbnails
+ For more information, see [#1194](https://github.com/GetStream/stream-chat-swift/pull/1194)
 - `mentionText(for:)` function added to `ComposerVC` for customizing the text displayed for mentions [#1188](https://github.com/GetStream/stream-chat-swift/issues/1188) [#1000](https://github.com/GetStream/stream-chat-swift/issues/1000)
 - `score` to `ChatMessageReactionData` so a slack-like reaction view is achievable. This would be used as content in `ChatMessageReactionsView` [#1200](https://github.com/GetStream/stream-chat-swift/issues/1200)
 - Ability to send silent messages. Silent messages are normal messages with an additional `isSilent` value set to `true`. Silent messages don’t trigger push notification for the recipient.[#1211](https://github.com/GetStream/stream-chat-swift/pull/1211)
 - Expose `cid` on `Message` [#1215](https://github.com/GetStream/stream-chat-swift/issues/1215)
+- `showMediaPicker`/`showFilePicker`/`attachmentsPickerActions` functions added to `ComposerVC` so it's possible to customize media/document pickers and add extend action sheet with actions for custom attachment types [#1194](https://github.com/GetStream/stream-chat-swift/pull/1194)
 
 ### 🔄 Changed
 - `scrollToLatestMessageButton` is now visible every time the last message is not visible. Not only when there is unread message. [#1208](https://github.com/GetStream/stream-chat-swift/pull/1208) 
+- `mediaPickerVC` in `ComposerVC` updated to show both photos and videos [#1194](https://github.com/GetStream/stream-chat-swift/pull/1194)
 
 ### 🐞 Fixed 
 - Fix sorting Member List by `createdAt` causing an issue [#1185](https://github.com/GetStream/stream-chat-swift/issues/1185)
@@ -40,6 +56,8 @@ _June 23, 2021_
 - Fix composer overlapping last message. This happened for channels with typing events disabled. [#1210](https://github.com/GetStream/stream-chat-swift/issues/1210)
 - Fix an issue where composer textView's caret jumps to the end of input [#1117](https://github.com/GetStream/stream-chat-swift/issues/1117)
 - Fix deadlock in Controllers when `synchronize` is called in a delegate callback [#1214](https://github.com/GetStream/stream-chat-swift/issues/1214)
+- Fix restart uploading action not being propagated [#1194](https://github.com/GetStream/stream-chat-swift/pull/1194)
+- Fix uploading progress not visible on image uploading overlay [#1194](https://github.com/GetStream/stream-chat-swift/pull/1194)
 
 # [4.0.0-beta.3](https://github.com/GetStream/stream-chat-swift/releases/tag/4.0.0-beta.3)
 _June 11, 2021_